Add comprehensive tests for decimal buying costs in inventory management
- Introduced a new test suite to validate the handling of decimal unit costs in various inventory-related tables.
- Verified that batch buying costs, movement costs, and invoice totals support decimal rates while maintaining whole-KES values for original and remaining costs.
- Ensured that stock adjustments and sales processes correctly handle decimal costs without introducing rounding errors.
- Included tests for VAT-inclusive and exclusive purchases to confirm accurate tax calculations alongside decimal pricing.
- Validated that voiding sales and applying workbook updates maintain integrity of decimal cost allocations.

+++ supabase/tests/database/0042_purchase_intelligence.test.sql
@@ -47,7 +47,7 @@ select is((select price from public.product_variants
 select is((select count(*)::int from public.purchase_lines
   where purchase_id=(select purchase_id from pi_purchase)),1,'purchase line is retained');
 select is((select average_unit_cost from public.supplier_variant_performance
-  where supplier_id='42000000-0000-0000-0000-000000000003'),9000::bigint,
+  where supplier_id='42000000-0000-0000-0000-000000000003'),9000::numeric,
   'supplier product cost aggregate is accurate');
 
 select public.record_purchase_with_prices('42000000-0000-0000-0000-000000000004',
